Converted from doubles to floats

+++ src/cameras/Camera.h
@@ -24,15 +24,15 @@ public:
 
     /// @brief A rotation of the camera about its forward axis.
     /// @param angle rotation angle in degrees
-    void roll(const double angle);
+    void roll(const float angle);
 
     /// @brief Rotate the camera about the horizontal axis.
     /// @param angle rotation angle in degrees
-    void tilt(const double angle);
+    void tilt(const float angle);
 
     /// @brief Rotate the camera about its vertical axis.
     /// @param angle rotation angle in degrees
-    void pan(const double angle);
+    void pan(const float angle);
     //@}
 
     //@{
@@ -40,85 +40,85 @@ public:
 
     /// @brief Translation of the camera along its forward axis
     /// @param value the amount to move the camera
-    void dolly(const double value);
+    void dolly(const float value);
 
     /// @brief Translation of the camera along its vertical axis.
     /// @param value the amount to move the camera
-    void boom(const double value);
+    void boom(const float value);
     //@}
 
     //@{
     /// Get the normalized forward, horizontal, and vertical axes.
-    glm::dvec3 getForwardAxis();
-    glm::dvec3 getHorizontalAxis();
-    glm::dvec3 getVerticalAxis();
+    glm::vec3 getForwardAxis();
+    glm::vec3 getHorizontalAxis();
+    glm::vec3 getVerticalAxis();
     //@}
 
     //@{
     /// @brief Set/Get the position of the camera in object space.
     /// @param position position of the camera
-    void setPosition(const glm::dvec3 &position);
-    glm::dvec3 getPosition() const { return m_position; }
+    void setPosition(const glm::vec3 &position);
+    glm::vec3 getPosition() const { return m_position; }
     //@}
 
     /// @brief Get the position of the camera in world space.
     /// @return the position of the camera in world space
-    glm::dvec3 getWorldPosition() const;
+    glm::vec3 getWorldPosition() const;
 
     //@{
     /// @brief Set/Get the focal point of the camera in world coordinates. The default
     ///        focal point is the origin.
     /// @param focalPoint the focal point of the camera
-    void setFocalPoint(const glm::dvec3 &focalPoint);
-    glm::dvec3 getFocalPoint() const { return m_focalPoint; }
+    void setFocalPoint(const glm::vec3 &focalPoint);
+    glm::vec3 getFocalPoint() const { return m_focalPoint; }
     //@}
 
     //@{
     /// Set/get the camera view up direction. The view up vector will be
     /// normalized.
-    void setViewUp(const glm::dvec3 &up);
-    glm::dvec3 getViewUp() const { return m_viewUp; }
+    void setViewUp(const glm::vec3 &up);
+    glm::vec3 getViewUp() const { return m_viewUp; }
     //@}
 
     //@{
     /// @brief Set/Get the background color of the camera.
     /// @param color the background color
-    void setBackgroundColor(const Color3d &color) { m_background = color; }
-    Color3d getBackgroundColor() const { return m_background; }
+    void setBackgroundColor(const Color3f &color) { m_background = color; }
+    Color3f getBackgroundColor() const { return m_background; }
     //@}
 
     //@{
     /// @brief Set/Get the aperature radius of the camera.
     /// @param radius the aperature radius
-    void setAperatureRadius(const double radius) { m_aperatureRadius = radius; }
-    double getAperatureRadius() const { return m_aperatureRadius; }
+    void setAperatureRadius(const float radius) { m_aperatureRadius = radius; }
+    float getAperatureRadius() const { return m_aperatureRadius; }
     //@}
 
     /// @brief The viewing transformation for moving objects from world space to view space.
     /// @return world to view transformation matrix
-    glm::dmat4 getViewMatrix();
+    glm::mat4 getViewMatrix();
 
     //@{
     /// Set/get the camera to world matrix for moving the camera to world space.
-    glm::dmat4 getCameraToWorldMatrix() const { return m_modelMatrix; }
-    void setCameraToWorldMatrix(const glm::dmat4 &matrix);
+    glm::mat4 getCameraToWorldMatrix() const { return m_modelMatrix; }
+    void setCameraToWorldMatrix(const glm::mat4 &matrix);
     //@}
 
     /// @brief Generate a Halton sequence of points in the range [0,1).
     /// @param N the number of points to generate
     /// @param b the base of the sequence
     /// @param out the output array of points
-    static void generateHaltonSequence(const int N, const int b, double *out);
+    static void generateHaltonSequence(const int N, const int b, float *out);
 
 private:
-    glm::dvec3 m_position;
-    glm::dvec3 m_focalPoint;
-    glm::dvec3 m_viewUp;
+    glm::vec3 m_position;
+    glm::vec3 m_focalPoint;
+    glm::vec3 m_viewUp;
 
-    Color3d m_background;
+    Color3f m_background;
 
-    double     m_aperatureRadius;
-    glm::dmat4 m_modelMatrix;
+    float     m_aperatureRadius;
+    glm::mat4 m_modelMatrix;
 };
 } // namespace raytracer
