+++ src/cameras/OrthographicCamera.cpp
@@ -7,9 +7,9 @@ namespace raytracer
 //----------------------------------------------------------------------------------
 OrthographicCamera::OrthographicCamera(int width,
                                        int height,
-                                       double fovDistance,
-                                       double near,
-                                       double far) :
+                                       float fovDistance,
+                                       float near,
+                                       float far) :
     m_width(width),
     m_height(height),
     m_scale(1.0),
@@ -18,17 +18,17 @@ OrthographicCamera::OrthographicCamera(int width,
     m_near(near),
     m_far(far)
 {
-    double widthd = static_cast<double>(m_width);
-    double heightd = static_cast<double>(m_height);
-    double aspect =  widthd / heightd;
+    float widthd = static_cast<float>(m_width);
+    float heightd = static_cast<float>(m_height);
+    float aspect =  widthd / heightd;
 
-    double w = widthd * aspect * m_scale * 0.5;
-    double h = m_scale * heightd * 0.5;
+    float w = widthd * aspect * m_scale * 0.5;
+    float h = m_scale * heightd * 0.5;
 
-    double xmin = -1.0 * w;
-    double xmax = w;
-    double ymin = -1.0 * h;
-    double ymax = h;
+    float xmin = -1.0 * w;
+    float xmax = w;
+    float ymin = -1.0 * h;
+    float ymax = h;
 
     m_orthographicMatrix = glm::ortho(xmin, xmax, ymin, ymax, near, far);
 }
@@ -39,29 +39,29 @@ void OrthographicCamera::reset()
     // TODO: save original values to reset to previous values
     this->Camera::reset();
 
-    m_scale = 1.0;
-    m_fovDistance = 100.0;
-    m_near = 0.1;
-    m_far = 1000.0;
+    m_scale = 1.0f;
+    m_fovDistance = 100.0f;
+    m_near = 0.1f;
+    m_far = 1000.0f;
 }
 
 //----------------------------------------------------------------------------------
-void OrthographicCamera::setClippingRange(const double near, const double far)
+void OrthographicCamera::setClippingRange(const float near, const float far)
 {
     m_near = near;
     m_far = far;
 
-    double widthf = static_cast<double>(m_width);
-    double heightf = static_cast<double>(m_height);
-    double aspect =  widthf / heightf;
+    float widthf = static_cast<float>(m_width);
+    float heightf = static_cast<float>(m_height);
+    float aspect =  widthf / heightf;
 
-    double w = widthf * aspect * m_scale * 0.5;
-    double h = m_scale * heightf * 0.5;
+    float w = widthf * aspect * m_scale * 0.5f;
+    float h = m_scale * heightf * 0.5f;
 
-    double xmin = w * -1.0;
-    double xmax = w;
-    double ymin = h * -1.0;
-    double ymax = h;
+    float xmin = w * -1.0f;
+    float xmax = w;
+    float ymin = h * -1.0f;
+    float ymax = h;
 
     m_orthographicMatrix = glm::ortho(xmin, xmax, ymin, ymax, m_near, m_far);
 }
@@ -72,51 +72,51 @@ void OrthographicCamera::setScreenSize(const int width, const int height)
     m_width = width;
     m_height = height;
 
-    double widthf = static_cast<double>(m_width);
-    double heightf = static_cast<double>(m_height);
-    double aspect =  widthf / heightf;
+    float widthf = static_cast<float>(m_width);
+    float heightf = static_cast<float>(m_height);
+    float aspect =  widthf / heightf;
 
-    double w = widthf * aspect * m_scale * 0.5;
-    double h = m_scale * heightf * 0.5;
+    float w = widthf * aspect * m_scale * 0.5f;
+    float h = m_scale * heightf * 0.5f;
 
-    double xmin = w * -1.0;
-    double xmax = w;
-    double ymin = h * -1.0;
-    double ymax = h;
+    float xmin = w * -1.0f;
+    float xmax = w;
+    float ymin = h * -1.0f;
+    float ymax = h;
 
     m_orthographicMatrix = glm::ortho(xmin, xmax, ymin, ymax, m_near, m_far);
 }
 
 //----------------------------------------------------------------------------------
-void OrthographicCamera::zoom(const double factor)
+void OrthographicCamera::zoom(const float factor)
 {
     m_zoomFactor = factor;
     m_scale /= factor;
 
-    double widthf = static_cast<double>(m_width);
-    double heightf = static_cast<double>(m_height);
-    double aspect =  widthf / heightf;
+    float widthf = static_cast<float>(m_width);
+    float heightf = static_cast<float>(m_height);
+    float aspect =  widthf / heightf;
 
-    double w = widthf * aspect * m_scale * 0.5;
-    double h = m_scale * heightf * 0.5;
+    float w = widthf * aspect * m_scale * 0.5;
+    float h = m_scale * heightf * 0.5;
 
-    double xmin = w * -1.0;
-    double xmax = w;
-    double ymin = h * -1.0;
-    double ymax = h;
+    float xmin = w * -1.0;
+    float xmax = w;
+    float ymin = h * -1.0;
+    float ymax = h;
 
     m_orthographicMatrix = glm::ortho(xmin, xmax, ymin, ymax, m_near, m_far);
 }
 
 //----------------------------------------------------------------------------------
 Ray *
-OrthographicCamera::generateRay(const glm::dvec2 &pixel)
+OrthographicCamera::generateRay(const glm::vec2 &pixel)
 {
     auto p = pixel * m_fovDistance;
-    auto origin = glm::dvec3(p, 0.0);
+    auto origin = glm::vec3(p, 0.0f);
 
     Ray *ray = new Ray(origin,
-                       glm::dvec3(0.0, 0.0, -1.0),
+                       glm::vec3(0.0f, 0.0f, -1.0f),
                        std::abs(m_near),
                        std::abs(m_far));
 
@@ -132,10 +132,10 @@ OrthographicCamera::copy(const ProjectionCamera * const camera)
         // Projection Camera
         this->zoom(camera->getZoomFactor());
 
-        glm::dvec2 clippingRange = camera->getClippingRange();
+        glm::vec2 clippingRange = camera->getClippingRange();
         this->setClippingRange(clippingRange[0], clippingRange[1]);
 
-        glm::dvec2 screenSize = camera->getScreenSize();
+        glm::vec2 screenSize = camera->getScreenSize();
         this->setScreenSize(screenSize.x, screenSize.y);
 
         // Camera
